requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.7.0
huggingface-hub>=0.30.0,<1.0
transformers==4.36.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
pypdf2==3.0.1
docx2txt==0.8
plotly==5.17.0
wordcloud==1.9.2
textstat==0.7.3
beautifulsoup4==4.12.2
markdown==3.5.1
chromadb==0.4.18
tiktoken==0.5.2
streamlit-authenticator==0.2.3
streamlit-extras==0.3.5
python-dateutil==2.8.2
regex==2023.10.3
pytz==2023.3